+++ src/main/java/com/backend/backend/models/HistorialCompra.java
@@ -3,6 +3,7 @@ package com.backend.backend.models;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -24,5 +25,5 @@ public class HistorialCompra {
     private Date fechaCompra;
 
     @Column(name = "importe", nullable = false)
-    private Double importe;
+    private BigDecimal importe;
 }
